Test createSubmission directly, replacing the HTTP-level /submit tests
Now that the submission-creation logic lives in createSubmission.js, the
tests can call it directly with a stubbed saveUser instead of going
through the express app. The same six behaviors are covered, but the
fetch/FormData/forceSsl/env-var/app.listen workarounds are gone, since
the module no longer needs server.js's import-time quirks.
