--- src/server.test.js
@@ -1,352 +1,0 @@
-/**
- * @jest-environment node
- *
- * Characterization tests for server.js's /submit route: real HTTP through
- * the actual express app, against a real Parse Server 2.8.4 backed by a real
- * MongoDB 4.4 started in-memory by mongodb-memory-server, so the full
- * multer/express/Parse path is exercised. These lock in the handler's
- * current behavior so the submission-creation logic can later be extracted
- * from it without changing what /submit does.
- */
-
-import http from 'http';
-import net from 'net';
-
-import Parse from 'parse/node';
-
-const { MongoMemoryServer } = require('mongodb-memory-server');
-// parse-server is deliberately installed on demand instead of being a project
-// dependency (see jest.globalSetup.js)
-const { ParseServer } = require('parse-server');
-
-// chunk-manifest.json is generated by the webpack build and doesn't exist
-// in the test environment; server.js only needs it to exist at import time.
-jest.mock('./chunk-manifest.json', () => ({}), { virtual: true });
-
-// parse-server skips its cloud/URL verification (and test-unfriendly process
-// listeners) when TESTING is set. Must be set before it is constructed.
-process.env.TESTING = '1';
-
-jest.setTimeout(30000);
-
-const email = 'test@example.com';
-
-// The /submit handler mutates process.env.TZ (and only restores it on the
-// future-timestamp error path), so restore it between tests to keep them
-// independent.
-const originalTZ = process.env.TZ;
-
-// Minimal 1x1 PNG and a minimal mp4 `ftyp` box, both small enough to inline
-// while still being recognized by the file-type detector.
-const pngBuffer = Buffer.from(
-  'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mP8z8BQDwAEhQGAhKmMIQAAAABJRU5ErkJggg==',
-  'base64',
-);
-const mp4Buffer = Buffer.concat([
-  Buffer.from([0, 0, 0, 0x18]), // box size
-  Buffer.from('ftypisom'), // ftyp + major brand isom
-  Buffer.from([0, 0, 2, 0]),
-  Buffer.from('isomiso2'),
-]);
-
-// In-memory stand-in for a files adapter: the submission logic under test
-// only cares that Parse.File#save() succeeds and produces a URL, not where
-// the bytes are stored, and the real GridStore adapter opens its own MongoDB
-// connection that nothing shuts down (keeping the test process alive after
-// the run).
-class InMemoryFilesAdapter {
-  constructor() {
-    this.files = new Map();
-  }
-
-  createFile(filename, data) {
-    this.files.set(filename, data);
-    return Promise.resolve();
-  }
-
-  deleteFile(filename) {
-    this.files.delete(filename);
-    return Promise.resolve();
-  }
-
-  getFileData(filename) {
-    return Promise.resolve(this.files.get(filename));
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  getFileLocation(config, filename) {
-    return `${config.mount}/files/${config.applicationId}/${encodeURIComponent(filename)}`;
-  }
-}
-
-describe('/submit', () => {
-  let mongo;
-  let parseServer;
-  let httpServer;
-  let baseUrl;
-  let user;
-
-  beforeAll(async () => {
-    // MongoDB 4.4 is the newest version whose wire protocol parse-server
-    // 2.8.4's bundled mongodb driver can talk to. Note: the binary must be
-    // downloaded once (mongodb-memory-server caches it), and on Ubuntu 24+
-    // mongod 4.4 needs libssl1.1 installed.
-    mongo = await MongoMemoryServer.create({ binary: { version: '4.4.14' } });
-
-    // create() can resolve a moment before mongod accepts connections;
-    // parse-server connects to it in its constructor, so wait until the
-    // port actually accepts a TCP connection.
-    const { port } = new URL(mongo.getUri());
-    const attemptConnection = (resolve, reject, attempts) => {
-      const socket = net.connect(Number(port), '127.0.0.1');
-      socket.once('connect', () => {
-        socket.destroy();
-        resolve();
-      });
-      socket.once('error', () => {
-        socket.destroy();
-        if (attempts < 100) {
-          setTimeout(
-            () => attemptConnection(resolve, reject, attempts + 1),
-            50,
-          );
-        } else {
-          reject(new Error('mongod never accepted connections'));
-        }
-      });
-    };
-    await new Promise((resolve, reject) => {
-      attemptConnection(resolve, reject, 0);
-    });
-
-    parseServer = ParseServer.start(
-      {
-        databaseURI: mongo.getUri(),
-        appId: 'test-app',
-        masterKey: 'test-master',
-        // Only used as a placeholder; the client points at the real port.
-        serverURL: 'http://localhost/parse',
-        mountPath: '/parse',
-        port: 0,
-        verbose: false,
-        // server.js's logIn() throws its "check your email" error for users
-        // without emailVerified, which the test server would otherwise
-        // leave false after signup.
-        verifyUserEmails: false,
-        filesAdapter: new InMemoryFilesAdapter(),
-      },
-      () => {},
-    );
-    await new Promise((resolve, reject) => {
-      parseServer.server.once('listening', resolve);
-      parseServer.server.once('error', reject);
-    });
-
-    // server.js reads its config from env vars at import time, so point
-    // them at the test server before requiring it. (The dotenv preload
-    // doesn't override variables that are already set.)
-    process.env.PARSE_APP_ID = 'test-app';
-    process.env.PARSE_JAVASCRIPT_KEY = 'test-js-key';
-    process.env.PARSE_MASTER_KEY = 'test-master';
-    process.env.PARSE_SERVER_URL = `http://localhost:${parseServer.server.address().port}/parse`;
-    process.env.HEROKU_RELEASE_VERSION = 'v123';
-
-    // server.js calls app.listen(config.port) in a deferred microtask at
-    // import time, which would bind a real port and leak the handle.
-    // Replace listen (synchronously, before that microtask runs) with one
-    // that listens on port 0 and captures the server.
-    // eslint-disable-next-line global-require
-    const app = require('./server.js').default;
-    let listeningResolve;
-    const listeningPromise = new Promise(resolve => {
-      listeningResolve = resolve;
-    });
-    app.listen = (listenPort, callback) => {
-      httpServer = http.createServer(app);
-      httpServer.listen(0, () => {
-        listeningResolve();
-        if (callback) callback();
-      });
-      return httpServer;
-    };
-    await listeningPromise;
-    baseUrl = `http://127.0.0.1:${httpServer.address().port}`;
-
-    // server.js's import initialized our shared Parse client, so the test
-    // user can be created directly against the test server. Parse Server
-    // 2.8.4 leaves emailVerified unset even with verifyUserEmails: false,
-    // and server.js's logIn() throws its "check your email" error for
-    // unverified users, so mark the user verified explicitly.
-    user = new Parse.User();
-    user.setUsername(email);
-    user.setPassword('test-password');
-    await user.signUp();
-    user.set('emailVerified', true);
-    await user.save(null, { sessionToken: user.getSessionToken() });
-  });
-
-  afterAll(async () => {
-    // Guard each step so a beforeAll failure doesn't cascade into teardown.
-    if (httpServer) {
-      await new Promise(resolve => httpServer.close(resolve));
-    }
-    if (parseServer) {
-      await new Promise(resolve => parseServer.server.close(resolve));
-      parseServer.handleShutdown();
-    }
-    if (mongo) {
-      await mongo.stop();
-    }
-  });
-
-  afterEach(() => {
-    process.env.TZ = originalTZ;
-  });
-
-  const validFields = overrides => ({
-    email,
-    password: 'test-password',
-    FirstName: 'Test',
-    LastName: 'User',
-    Phone: '5551234567',
-    testify: 'true',
-    plate: 'ABC1234',
-    licenseState: 'NY',
-    typeofreport: 'complaint',
-    typeofcomplaint: 'Blocked the bike lane',
-    reportDescription: 'Blocking the lane',
-    can_be_shared_publicly: 'true',
-    latitude: '40.7128',
-    longitude: '-74.006',
-    formatted_address: 'New York, NY',
-    CreateDate: new Date(Date.now() - 60 * 60 * 1000).toISOString(),
-    ...overrides,
-  });
-
-  const postSubmission = async ({
-    fields = validFields(),
-    files = [],
-  } = {}) => {
-    const formData = new FormData();
-    Object.entries(fields).forEach(([key, value]) => {
-      formData.append(key, String(value));
-    });
-    files.forEach(({ buffer, filename }) => {
-      formData.append('attachmentData[]', new Blob([buffer]), filename);
-    });
-    const response = await fetch(`${baseUrl}/submit`, {
-      method: 'POST',
-      // force-ssl-heroku redirects requests that aren't https
-      headers: { 'X-Forwarded-Proto': 'https' },
-      body: formData,
-    });
-    const body = await response.json();
-    return { status: response.status, body };
-  };
-
-  test('saves the form fields onto the submission, ACLed to the user', async () => {
-    const { status, body } = await postSubmission();
-
-    expect(status).toBe(200);
-    expect(body.submission).toMatchObject({
-      FirstName: 'Test',
-      LastName: 'User',
-      Phone: '5551234567',
-      testify: true,
-      Username: email,
-      typeofreport: 'complaint',
-      selectedReport: 1,
-      colorTaxi: 'Black',
-      medallionNo: 'ABC1234',
-      license: 'ABC1234',
-      state: 'NY',
-      typeofcomplaint: 'Blocked the bike lane',
-      passenger: false,
-      locationNumber: 1,
-      latitude: '40.7128',
-      longitude: '-74.006',
-      latitude1: 40.7128,
-      longitude1: -74.006,
-      loc1_address: 'New York, NY',
-      reportDescription: 'Blocking the lane',
-      can_be_shared_publicly: true,
-      status: 0,
-      operating_system: 'web',
-      version_number: 123,
-      reqnumber: 'N/A until submitted to 311',
-    });
-    expect(body.submission.objectId).toBeDefined();
-    expect(body.submission.user).toMatchObject({
-      __type: 'Object',
-      className: '_User',
-      objectId: user.id,
-    });
-    expect(body.submission.timeofreport).toMatch(/^\d{4}-\d{2}-\d{2}T/);
-  });
-
-  test('marks non-complaint reports differently via selectedReport', async () => {
-    const { status, body } = await postSubmission({
-      fields: validFields({ typeofreport: 'compliment' }),
-    });
-
-    expect(status).toBe(200);
-    expect(body.submission.selectedReport).toBe(0);
-  });
-
-  test('rejects submissions missing a required field', async () => {
-    const { status, body } = await postSubmission({
-      fields: validFields({ plate: '' }),
-    });
-
-    expect(status).toBe(500);
-    expect(body.error.message).toBe('plate is required');
-  });
-
-  test('rejects timestamps in the future, restoring the previous timezone', async () => {
-    const { status, body } = await postSubmission({
-      fields: validFields({
-        CreateDate: new Date(Date.now() + 60 * 60 * 1000).toISOString(),
-      }),
-    });
-
-    expect(status).toBe(500);
-    expect(body.error.message).toMatch(/^Timestamp cannot be in the future/);
-    expect(process.env.TZ).toBe(originalTZ);
-  });
-
-  test('uploads images as Parse Files under photoDataN and videos as URLs under videoDataN', async () => {
-    const { status, body } = await postSubmission({
-      files: [
-        { buffer: pngBuffer, filename: 'photo.png' },
-        { buffer: mp4Buffer, filename: 'video.mp4' },
-      ],
-    });
-
-    expect(status).toBe(200);
-    // The files controller prefixes stored file names with a hash, so match
-    // the suffix instead of the exact name.
-    expect(body.submission.photoData0).toMatchObject({
-      __type: 'File',
-      name: expect.stringMatching(/_photoData0\.png$/),
-    });
-    expect(body.submission.videoData0).toEqual(
-      expect.stringContaining('videoData0.mp4'),
-    );
-  });
-
-  test('keeps only the first 3 images (does not save 4th as videoData0)', async () => {
-    const { status, body } = await postSubmission({
-      files: [1, 2, 3, 4].map(() => ({
-        buffer: pngBuffer,
-        filename: 'photo.png',
-      })),
-    });
-
-    expect(status).toBe(200);
-    expect(body.submission.photoData0).toBeDefined();
-    expect(body.submission.photoData1).toBeDefined();
-    expect(body.submission.photoData2).toBeDefined();
-    expect(body.submission.photoData3).toBeUndefined();
-    expect(body.submission.videoData0).toBeUndefined();
-  });
-});
